requirements.txt
django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2-binary==2.8.3
python-decouple==3.1
python-dotenv==0.10.3
